Add files via upload
It is a project titled Gender disparity in Nobel prize winners. Data cleaning was done in power query after which the dta was loaded into power bi for visualization.
It was a project carried out by my team during an online training with Tech4Dev and I appreciate everyone that made this project a success.